Move Barelink work to new network plugin
